Update incubating and incubated project lists
Added incubating entries for enterprise-gateway and nb2kg.
Moved incubating entries for kernel-gateway, dashboards, traittypes,
and declarativewidgets to the inclubated section.  The last two of
which aren't actually in the jupyter GitHub, but in jupyter-widgets
and jupyter-attic, respectively.

Not sure if jupyter-attic entries should be included in the incubated
section.